test(inspector): capture probe sensors before Step, not after
The 100-tick headless inspector JSONL attributed sensors to the tick
that had just finished staging the *next* sense. clamped then disagreed
with the brain-facing vector from stage_sense (bd-16g.4.2).

- Clone the probe snapshot before submit Step so attribution matches
  the completed boundary.
- Closed-world fixture zeros food growth/decay/diffusion/respawn and
  initial_food so the 100-tick run cannot nibble ambient food.
- Control-plane food_growth_rate compare uses f32 EPSILON against the
  config field type; an f64 EPSILON made a rejected patch look like a
  mutation.

The JSONL dump under tests/artifacts is a local run product and is not
committed.

+++ crates/scriptbots-app/src/control.rs
@@ -1894,7 +1894,7 @@ pub(crate) mod tests {
             .and_then(Value::as_f64)
             .expect("food_growth_rate");
         assert!(
-            (value - f64::from(ScriptBotsConfig::default().food_growth_rate)).abs() < f64::EPSILON,
+            (value as f32 - ScriptBotsConfig::default().food_growth_rate).abs() < f32::EPSILON,
             "rejected update changed food_growth_rate to {value}"
         );
     }
